Fixed movement and Limited door click to occupied cell
Fixed cells dictionary so that movement was corrected. Col 0 in EAS and WES were flipped. Player can now leave the cells if they were trapped on the edges.

Limited doors that can be clicked to just be the doors that exist in the current cell

Added logging bool and function so I can turn off when I'm tired of seeing my mistakes